DBFS : améliorations diverses.
Chaque PDB peut donc posséder son propre DBFS, tous les scripts ont été
adaptés pour.

Le service CRS est crée, mais il y a quelques bugs :

* L'arrêt de la base n'est possible qu'avec l'option `-force`

* Le service du DBFS ne démarre pas après un arrêt démarrage de la base.

Scripts impactés par la mise en oeuvre d'un DBFS
    db/dbfs/configure_fuse_and_dbfs_mount_point.sh
    db/dbfs/create_crs_resource_for_dbfs.sh
    db/dbfs/create_dbfs.sh
    db/dbfs/create_dbfs.sql
    db/dbfs/drop_all.sh
    db/dbfs/readme.md
    wallet/create_wallet.sh
    dblib.sh
    db/dbfs/mount-dbfs
    db/dbfs/root_drop_all.sh

Test espace disque suffisant avant de lancer le clonage.
    cfglib.sh
    database_servers/clone_master.sh

database_servers/clean_up_infra.sh
    Corrections/évolutions.

database_servers/revert_to_master.sh
    Corrections.

Correction bug.
    gilib.sh
    ntp/configure_ntp.sh
    ntp/readme.md
    san/check_target.sh
    setup_first_vms/03_setup_infra_vm.sh
    ssh/setup_ssh_equivalence.sh
    database_servers/crontab_workaround_ntp.txt
    database_servers/sync_ntp.sh
    ntp/force_sync_ntp.sh

virtualbox/old/compile_guest_additions.sh
    Test, pour futur évolution.

+++ database_servers/clean_up_infra.sh
@@ -9,12 +9,18 @@ PLELIB_OUTPUT=FILE
 EXEC_CMD_ACTION=EXEC
 
 typeset -r ME=$0
-typeset -r str_usage="Usage : $ME -db=<str> -delete_vms=yes"
+typeset -r str_usage="
+Usage : $ME
+	-db=<str>
+	[-keep_vm]         keep VMs.
+	[-keep_cfg_files]  keep configuration files.
+"
 
 script_banner $ME $*
 
 typeset	db=undef
 typeset	delete_vms=yes
+typeset	remove_cfg_file=yes
 
 while [ $# -ne 0 ]
 do
@@ -29,8 +35,13 @@ do
 			shift
 			;;
 
-		-delete_vms=*)
-			delete_vms=${1##*=}
+		-keep_vms)
+			delete_vms=no
+			shift
+			;;
+
+		-keep_cfg_file)
+			remove_cfg_file=no
 			shift
 			;;
 
@@ -65,6 +76,8 @@ then
 	do
 		cfg_load_node_info $db $inode
 		remove_from_known_hosts $cfg_server_name
+		exec_cmd "ssh -t $dns_conn					\
+				plescripts/ssh/remove_from_known_host.sh -host=$cfg_server_name"
 		if [[ $max_nodes -gt 1 && $inode -eq 1 ]]
 		then	# Supprime les scans.
 			remove_from_known_hosts ${db}-scan
@@ -80,7 +93,7 @@ exec_cmd "ssh -t $dns_conn plescripts/dns/remove_db_from_dns.sh -db=$db"
 LN
 
 line_separator
-info "Remove keys from DNS"
+info "Remove keys"
 exec_cmd "ssh -t $dns_conn plescripts/dns/clean_up_ssh_authorized_keys.sh"
 LN
 
@@ -94,7 +107,10 @@ info "Clean up local DNS cache :"
 exec_cmd -c sudo systemctl restart nscd.service
 LN
 
-line_separator
-info "Remove $cfg_path_prefix/$db"
-exec_cmd -c "rm -rf $cfg_path_prefix/$db"
-LN
+if [ $remove_cfg_file == yes ]
+then
+	line_separator
+	info "Remove $cfg_path_prefix/$db"
+	exec_cmd -c "rm -rf $cfg_path_prefix/$db"
+	LN
+fi
